move logic to template_expander
